Load Doctrine tasks only if Doctrine is installed
