+++ Addon/Doctrine/Task/EntityManager/DoctrineDetacherTask.php
@@ -8,38 +8,38 @@
 * file that was distributed with this source code.
 */
 
-namespace CleverAge\ProcessBundle\Task\Doctrine;
+namespace CleverAge\ProcessBundle\Addon\Doctrine\Task\EntityManager;
 
 use CleverAge\ProcessBundle\Model\ProcessState;
 use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManagerInterface;
 
 /**
- * Remove Doctrine entities
+ * Detach Doctrine entities from unit of work
  *
  * @author Valentin Clavreul <vclavreul@clever-age.com>
  * @author Vincent Chalnot <vchalnot@clever-age.com>
  */
-class DoctrineRemoverTask extends AbstractDoctrineTask
+class DoctrineDetacherTask extends AbstractDoctrineTask
 {
     /**
      * @param ProcessState $state
      *
      * @throws \UnexpectedValueException
      * @throws \Doctrine\ORM\ORMInvalidArgumentException
-     * @throws \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      * @throws \InvalidArgumentException
-     * @throws \Doctrine\ORM\ORMException
      */
     public function execute(ProcessState $state)
     {
         $entity = $state->getInput();
+        if (null === $entity) {
+            throw new \RuntimeException('DoctrineWriterTask does not allow null input');
+        }
         $class = ClassUtils::getClass($entity);
         $entityManager = $this->doctrine->getManagerForClass($class);
         if (!$entityManager instanceof EntityManagerInterface) {
             throw new \UnexpectedValueException("No manager found for class {$class}");
         }
-        $entityManager->remove($entity);
-        $entityManager->flush();
+        $entityManager->detach($entity);
     }
 }
