+++ Addon/Doctrine/Task/EntityManager/DoctrineRemoverTask.php
@@ -8,39 +8,36 @@
 * file that was distributed with this source code.
 */
 
-namespace CleverAge\ProcessBundle\Task\Doctrine;
+namespace CleverAge\ProcessBundle\Addon\Doctrine\Task\EntityManager;
 
 use CleverAge\ProcessBundle\Model\ProcessState;
 use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManagerInterface;
 
 /**
- * Detach Doctrine entities from unit of work
+ * Remove Doctrine entities
  *
  * @author Valentin Clavreul <vclavreul@clever-age.com>
  * @author Vincent Chalnot <vchalnot@clever-age.com>
  */
-class DoctrineDetacherTask extends AbstractDoctrineTask
+class DoctrineRemoverTask extends AbstractDoctrineTask
 {
     /**
      * @param ProcessState $state
      *
      * @throws \UnexpectedValueException
      * @throws \Doctrine\ORM\ORMInvalidArgumentException
-     * @throws \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      * @throws \InvalidArgumentException
      */
     public function execute(ProcessState $state)
     {
         $entity = $state->getInput();
-        if (null === $entity) {
-            throw new \RuntimeException('DoctrineWriterTask does not allow null input');
-        }
         $class = ClassUtils::getClass($entity);
         $entityManager = $this->doctrine->getManagerForClass($class);
         if (!$entityManager instanceof EntityManagerInterface) {
             throw new \UnexpectedValueException("No manager found for class {$class}");
         }
-        $entityManager->detach($entity);
+        $entityManager->remove($entity);
+        $entityManager->flush();
     }
 }
