+++ Resources/config/services-doctrine/task.yml
@@ -1,0 +1,8 @@
+services:
+    CleverAge\ProcessBundle\Addon\Doctrine\Task\:
+        resource: '../../../Addon/Doctrine/Task/*'
+        autowire: true
+        public: true
+        shared: false
+        tags:
+            - { name: monolog.logger, channel: cleverage_process_task }
